kurang tarik investasi

+++ database/migrations/2024_07_02_151420_create_anggotas_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->string('nama')->nullable();
             $table->string('telp')->nullable();
             $table->text('alamat')->nullable();
-            $table->decimal('poin')->nullable();
+            $table->bigInteger('poin')->nullable();
             $table->softDeletes();
             $table->timestamps();
         });
